LifeEventType を LifeEventParams ファイルに移動
