+++ lib/i18n/extensions/i18n_common.dart
@@ -1,6 +1,6 @@
 import 'package:intl/intl.dart';
 
-import '../../models/common/life_event.dart';
+import '../../models/common/life_event_params/life_event_params.dart';
 import '../i18n.dart';
 
 /// Common Locale Text
