+++ lib/models/common/life_event_params/gain_life_items_params.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../life_event.dart';
 import 'life_event_params.dart';
 import 'target_life_item_params.dart';
