+++ lib/models/common/life_event_params/nothing_params.dart
@@ -1,6 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../life_event.dart';
 import 'life_event_params.dart';
 
 part 'nothing_params.freezed.dart';
